+++ chbook.h
@@ -12,7 +12,7 @@
 typedef struct {
   char *vname, *units;
   double xmin, xmax, dx;
-  long xbins;
+  int32_t xbins;
   double *value, total;
   long length, count;
 } book1;
@@ -20,9 +20,9 @@ typedef struct {
 typedef struct {
   char *xname, *yname, *xunits, *yunits;
   double xmin, xmax, dx;
-  long xbins;
+  int32_t xbins;
   double ymin, ymax, dy;
-  long ybins;
+  int32_t ybins;
   double *value, total;
   long length, count;
 } book2;
@@ -31,7 +31,7 @@ typedef struct {
   char **vname, **units;
   long nD;
   double *xmin, *xmax, *dx;
-  long *xbins;
+  int32_t *xbins;
   double *value, total;
   long length, count;
 } ntuple;
@@ -40,28 +40,28 @@ typedef struct {
   char **vname, **units;
   double *xmin, *xmax, *dx;
   long nD;
-  long bins;
+  int32_t bins;
   double **value, total;
   long length, count;
 } book1m;
 
-book1 *chbook1(char *vName, char *units, double xmin, double xmax, long xbins);
+book1 *chbook1(char *vName, char *units, double xmin, double xmax, int32_t xbins);
 void chfill1(book1 *bName, double x, double Frequency);
 void free_hbook1(book1 *x);
 
 book2 *chbook2(char *xName, char *yName, char *xunits, char *yunits,
-               double xmin, double xmax, double ymin, double ymax, long xbins, long ybins);
+               double xmin, double xmax, double ymin, double ymax, int32_t xbins, int32_t ybins);
 void chfill2(book2 *bName, double x, double y, double Frequency);
 void free_hbook2(book2 *x);
 
-ntuple *chbookn(char **vName, char **units, long NDimension, double *xmin, double *xmax, long *xbins, long offset);
+ntuple *chbookn(char **vName, char **units, long NDimension, double *xmin, double *xmax, int32_t *xbins, long offset);
 void chfilln(ntuple *bName, double *x, double Frequency, long offset);
 void free_hbookn(ntuple *x);
 ntuple *readbookn(char *inputfile, long i_page);
 double interpolate_bookn(ntuple *bName, double *x0, double *x, long offset, long normalize);
 
-book1m *chbook1m(char **vName, char **units, double *xmin, double *xmax, long *xbins, long column_number);
-void chfill1m(book1m *bName, double *x, double Frequency, long *xbins, long column_number);
+book1m *chbook1m(char **vName, char **units, double *xmin, double *xmax, int32_t *xbins, long column_number);
+void chfill1m(book1m *bName, double *x, double Frequency, int32_t *xbins, long column_number);
 void free_hbook1m(book1m *x);
 
 void chprint1(book1 *bName, char *filename, char *description, SDDS_DEFINITION *parameter_definition,
